Fix two bugs about register and bind email

+++ fronted_v2/js/email_handlers.js
@@ -29,8 +29,11 @@ function bindEmail() {
                 const resData = error.responseJSON;
                 alert(JSON.stringify(resData.message));
                 console.log(error);
-                localStorage.clear();
-                parent.location.href="userLogin.html";
+
+                if (resData.errcode == 2) {
+                    localStorage.clear();
+                    parent.location.href="userLogin.html";
+                }
             }
         });
     }
